Split routes with express routers - site functionality remains the same
