+++ package.json
@@ -13,6 +13,7 @@
     "ejs": "^3.1.9",
     "ejs-mate": "^4.0.0",
     "express": "^4.18.2",
+    "express-session": "^1.17.3",
     "joi": "^17.11.0",
     "method-override": "^3.0.0",
     "mongoose": "^8.0.3",
